app/streamlit-app/requirements.txt
pyvespa # @ git+https://github.com/vespa-engine/pyvespa.git
sentence_transformers
streamlit
